sum_of_treatmnt
