--- Main.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-  <link rel="stylesheet" href="/static/Main.css">
-</head>
-<body>
-    <nav class="navMenu">
-        <a href=Login>Login</a>
-        <div class="dot"></div>
-    </nav>
-  </body>
-</html>
