--- Order_Doctor_P.html
@@ -1,54 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-        <link rel="stylesheet" href="/static/Style.css">
-    <style>
-        /* Chrome, Safari, Edge, Opera */
-        input::-webkit-outer-spin-button,
-        input::-webkit-inner-spin-button {
-          -webkit-appearance: none;
-          margin: 0;
-        }
-
-        /* Firefox */
-        input[type=number] {
-          -moz-appearance: textfield;
-        }
-        .table{
-            width: 60%;
-        }
-    </style>
-</head>
-<body>
-    <table class="table">
-       <tr>
-           {% for header in headings %}
-           <th>{{ header }}</th>
-           {% endfor %}
-       </tr>
-        {% for row in data %}
-        <tr>
-            {% for x in row %}
-            <td>{{ x }}</td>
-            {% endfor %}
-        </tr>
-        {% endfor %}
-    </table>
-    <div class="form">
-        <form action=Order_Doctor method="POST" class="form">
-            <label for="first_name"></label><input type="text" name="first_name" id="first_name" placeholder="First Name" autocomplete="off">
-            <label for="last_name"></label><input type="text" name="last_name" id="last_name" placeholder="Last Name" autocomplete="off">
-            <label for="email"></label><input type="text" name="email" id="email" placeholder="Email" autocomplete="off">
-            <p>
-                <button type="submit" name="Order" value="Order"> Order </button>
-                <button type="submit" name="Back" value="Back"> Back </button>
-            </p>
-        </form>
-        {% if message %}
-            <p>{{ message }}</p>
-        {% endif %}
-    </div>
-</body>
-</html>
