--- Orders_History_P.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-        <link rel="stylesheet" href="/static/Style.css">
-    <style>
-        /* Chrome, Safari, Edge, Opera */
-        input::-webkit-outer-spin-button,
-        input::-webkit-inner-spin-button {
-          -webkit-appearance: none;
-          margin: 0;
-        }
-
-        /* Firefox */
-        input[type=number] {
-          -moz-appearance: textfield;
-        }
-        .table{
-            width: 60%;
-        }
-    </style>
-</head>
-<body>
-    <table class="table">
-       <tr>
-           {% for header in headings %}
-           <th>{{ header }}</th>
-           {% endfor %}
-       </tr>
-        {% for row in data %}
-        <tr>
-            {% for x in row %}
-            <td>{{ x }}</td>
-            {% endfor %}
-        </tr>
-        {% endfor %}
-    </table>
-    <div class="form">
-        <form action=Orders_History method="POST" class="form">
-            <button type="submit" name="Back" value="Back"> Back </button>
-        </form>
-    </div>
-</body>
-</html>
