+++ Patient_R.html
@@ -5,22 +5,29 @@
 {% block body %}
   <h3 class='text-center text-muted mb-3'>Patient Requisites</h3>
   <p class='text-right'><a href='{{url_for("add_user")}}' class='btn btn-success '>+Add Appointmen</a></p>
+
   <table class='table table-bordered'>
     <thead>
-      <th>Requisite Number</th>
+      <th>SN</th>
+      <th>First Name</th>
+      <th>Last Name</th>
       <th>Email</th>
       <th>Approval Status</th>
       <th>Approve The Appointment</th>
       <th>Delete</th>
+
     </thead>
     <tbody>
       {% for row in datas %}
        <tr>
-          <td>{{loop.index}}</td>
-          <td>{{row.UNAME}}</td>
-          <td>{{row.CONTACT}}</td>
-          <td><a href='{{url_for("edit_user",uid=row.UID)}}' class='btn btn-primary' onclick='return confirm("Are You Sure you want to Approve The Appointment")'>Click To Approve</a></td>
-          <td><a href='{{url_for("delete_user",uid=row.UID)}}' class='btn btn-danger' onclick='return confirm("Are You Sure")'>Delete</a></td>
+            <td>{{loop.index}}</td>
+            <td>{{row.first_name}}</td>
+            <td>{{row.last_name}}</td>
+            <td>{{row.email_p}}</td>
+            <td>{{row.CONTACT}}</td>
+            <td><a href='{{url_for("edit_user",uid=row.UID)}}' class='btn btn-primary' onclick='return confirm("Are You Sure you want to Approve The Appointment")'>Click To Approve</a></td>
+            <td><a href='{{url_for("delete_user",uid=row.UID)}}' class='btn btn-danger' onclick='return confirm("Are You Sure")'>Delete</a></td>
+           <td><a href='{{url_for("edit_user2",uid=row.UID)}}' class='btn btn-primary' onclick='return confirm("456")'>456454</a></td>
 
         </tr>
       {%endfor%}
