--- Style_Admin.css
@@ -1,73 +1,0 @@
-body, html {
-    margin: 10px;
-    font-family: sans-serif;
-    /*background-color: lightblue;*/
-}
-
-/*body{*/
-/*	margin: 0;*/
-/*	padding: 0;*/
-/*	display: flex;*/
-/*	justify-content: center;*/
-/*	align-items: center;*/
-/*	min-height: 100vh;*/
-/*	font-family: 'Jost', sans-serif;*/
-/*	background: linear-gradient(to bottom, #0f0c29, #302b63, #24243e);*/
-/*}*/
-
-table, td, th {
-    border: 2px solid #aaa;
-}
-
-.table{
-    margin-left: auto;
-    margin-right: auto;
-}
-
-table {
-    border-collapse: separate;
-    width: 30%;
-}
-
-th {
-    height: 40px;
-    width: 50px;
-}
-
-td {
-    text-align: center;
-    padding: 5px;
-    width: 20%;
-    height: auto;
-}
-
-.form {
-    text-align: center;
-    margin-top: 25px;
-}
-
-.time {
-    text-align: center;
-    margin-top: 25px;
-}
-
-.image {
-    width: 70%;
-    height: 70%;
-}
-
-input {
-    width: 180px;
-}
-
-.wrapper {
-    font-family: Verdana;
-    font-size: 150%;
-    text-align: center;
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    min-height: 35em;
-    vertical-align: middle ;
-    align-items: center;
-}
