+++ add_user.html
@@ -3,16 +3,29 @@
 {% block body %}
   <div class='row'>
     <div class='col-5 mx-auto'>
-      <h3>Add user</h3><hr>
+      <h3>Add request</h3><hr>
       <form method='post' action='{{url_for("add_user")}}'>
+            <div class='form-group'>
+          <label>First Name</label>
+          <input  type='text'   placeholder="Egbaria" name='First_Name' required class='form-control'>
+        </div>
+
+        <div class='form-group'>
+          <label>Last Name</label>
+          <input  type='text'   placeholder="Muhamad" name='Last_Name' required class='form-control'>
+        </div>
+
         <div class='form-group'>
           <label>Email</label>
           <input  type='email'   placeholder="sophie@example.com" name='uname' required class='form-control'>
         </div>
-        <div class='form-group'>
 
+<!--   hidden value that send to the database the staus for the Approval     -->
+        <div class='form-group'>
           <input type='hidden' name='contact' required class='form-control' value='none'>
+          <input type='hidden' name='sum' required class='form-control' value='none'>
         </div>
+
         <input type='submit' value='Submit' class='btn btn-primary'>
       </form>
