+++ edit_user.html
@@ -7,7 +7,7 @@
       <form method='post' action='{{url_for("edit_user",uid=datas.UID)}}'>
         <div class='form-group'>
           <label>Email</label>
-          <input  type='email'   placeholder="sophie@example.com" name='uname' required class='form-control' value='{{datas.UNAME}}'>
+          <input  type='email'   placeholder="sophie@example.com" name='uname' required class='form-control' value='{{datas.email_p}}'>
 
         </div>
         <div class='form-group'>
