+++ edit_user2.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+{% extends 'layout.html' %}
+{% block body %}
+  <div class='row'>
+    <div class='col-5 mx-auto'>
+      <h3>treatment summary</h3><hr>
+      <form method='post' action='{{url_for("edit_user2",uid=datas.UID)}}'>
+        <div class='form-group'>
+          <label>write the here</label>
+          <textarea  name="sum" rows="10" cols="50" required class='form-control'></textarea>
+        </div>
+
+        <input type='submit' value='Submit' class='btn btn-primary'>
+      </form>
+    </div>
+  </div>
+{% endblock %}
